Move tick event listening for aspect effects to Title
